+++ decomp/SDK/src/dolphin/gx/GXTransform.c
@@ -1,8 +1,7 @@
-#include <macros.h>
-
 #include <dolphin/gx.h>
 #include <dolphin/mtx.h>
 #include <dolphin/os.h>
+#include <macros.h>
 
 #include "GXPrivate.h"
 
@@ -46,7 +45,7 @@ GXSetProjection (f32 mtx[4][4], GXProjectionType type)
 
     CHECK_GXBEGIN (0xCD, "GXSetProjection");
 
-    __GXData->projType = type;
+    __GXData->projType   = type;
     __GXData->projMtx[0] = mtx[0][0];
     __GXData->projMtx[2] = mtx[1][1];
     __GXData->projMtx[4] = mtx[2][2];
@@ -82,7 +81,7 @@ GXSetProjectionv (f32* ptr)
 
     CHECK_GXBEGIN (0x109, "GXSetProjectionv");
 
-    __GXData->projType = ptr[0];
+    __GXData->projType   = (u32)ptr[0];
     __GXData->projMtx[0] = ptr[1];
     __GXData->projMtx[1] = ptr[2];
     __GXData->projMtx[2] = ptr[3];
@@ -90,7 +89,7 @@ GXSetProjectionv (f32* ptr)
     __GXData->projMtx[4] = ptr[5];
     __GXData->projMtx[5] = ptr[6];
 
-    reg = 0x00061020;
+    reg                  = 0x00061020;
     GX_WRITE_U8 (0x10);
     GX_WRITE_U32 (reg);
     GX_WRITE_XF_REG_F (32, __GXData->projMtx[0]);
@@ -190,12 +189,12 @@ WriteMTXPS4x2 (register f32 mtx[2][4], register volatile f32* dest)
 #endif
 }
 
-#define GX_WRITE_MTX_ELEM(addr, value)                                                             \
-    do {                                                                                           \
-        f32 xfData = (value);                                                                      \
-        GX_WRITE_F32 (value);                                                                      \
-        VERIF_MTXLIGHT ((addr), *(u32*)&xfData);                                                   \
-    }                                                                                              \
+#define GX_WRITE_MTX_ELEM(addr, value)                                                         \
+    do {                                                                                       \
+        f32 xfData = (value);                                                                  \
+        GX_WRITE_F32 (value);                                                                  \
+        VERIF_MTXLIGHT ((addr), *(u32*)&xfData);                                               \
+    }                                                                                          \
     while (0)
 
 void
@@ -207,7 +206,7 @@ GXLoadPosMtxImm (f32 mtx[3][4], u32 id)
     CHECK_GXBEGIN (0x1D8, "GXLoadPosMtxImm");
 
     addr = id * 4;
-    reg = addr | 0xB0000;
+    reg  = addr | 0xB0000;
 
     GX_WRITE_U8 (0x10);
     GX_WRITE_U32 (reg);
@@ -230,13 +229,13 @@ GXLoadPosMtxImm (f32 mtx[3][4], u32 id)
 }
 
 // this one uses cmpwi instead of cmplwi for some reason
-#define SET_REG_FIELD_(line, reg, size, shift, val)                                                \
-    do {                                                                                           \
-        ASSERTMSGLINE (line,                                                                       \
-                       ((s32)(val) & ~((1 << (size)) - 1)) == 0,                                   \
-                       "GX Internal: Register field out of range");                                \
-        (reg) = ((u32)(reg) & ~(((1 << (size)) - 1) << (shift))) | ((u32)(val) << (shift));        \
-    }                                                                                              \
+#define SET_REG_FIELD_(line, reg, size, shift, val)                                            \
+    do {                                                                                       \
+        ASSERTMSGLINE (line,                                                                   \
+                       ((s32)(val) & ~((1 << (size)) - 1)) == 0,                               \
+                       "GX Internal: Register field out of range");                            \
+        (reg) = ((u32)(reg) & ~(((1 << (size)) - 1) << (shift))) | ((u32)(val) << (shift));    \
+    }                                                                                          \
     while (0)
 
 void
@@ -247,7 +246,7 @@ GXLoadPosMtxIndx (u16 mtx_indx, u32 id)
 
     CHECK_GXBEGIN (0x208, "GXLoadPosMtxIndx");
     offset = id * 4;
-    reg = 0;
+    reg    = 0;
     SET_REG_FIELD (0x20E, reg, 12, 0, offset);
     SET_REG_FIELD (0x20F, reg, 4, 12, 11);
     SET_REG_FIELD_ (0x210, reg, 16, 16, mtx_indx);
@@ -267,7 +266,7 @@ GXLoadNrmMtxImm (f32 mtx[3][4], u32 id)
     CHECK_GXBEGIN (0x229, "GXLoadNrmMtxImm");
 
     addr = id * 3 + 0x400;
-    reg = addr | 0x80000;
+    reg  = addr | 0x80000;
 
     GX_WRITE_U8 (0x10);
     GX_WRITE_U32 (reg);
@@ -295,7 +294,7 @@ GXLoadNrmMtxImm3x3 (f32 mtx[3][3], u32 id)
     CHECK_GXBEGIN (0x256, "GXLoadNrmMtxImm3x3");
 
     addr = id * 3 + 0x400;
-    reg = addr | 0x80000;
+    reg  = addr | 0x80000;
 
     GX_WRITE_U8 (0x10);
     GX_WRITE_U32 (reg);
@@ -322,7 +321,7 @@ GXLoadNrmMtxIndx3x3 (u16 mtx_indx, u32 id)
 
     CHECK_GXBEGIN (0x284, "GXLoadNrmMtxIndx3x3");
     offset = id * 3 + 0x400;
-    reg = 0;
+    reg    = 0;
     SET_REG_FIELD (0x28A, reg, 12, 0, offset);
     SET_REG_FIELD (0x28B, reg, 4, 12, 8);
     SET_REG_FIELD_ (0x28C, reg, 16, 16, mtx_indx);
@@ -359,8 +358,8 @@ GXLoadTexMtxImm (f32 mtx[][4], u32 id, GXTexMtxType type)
     {
         addr = id * 4;
     }
-    count = (type == GX_MTX2x4) ? 8 : 12;
-    reg = addr | ((count - 1) << 16);
+    count = (u32)((type == GX_MTX2x4) ? 8 : 12);
+    reg   = addr | ((count - 1) << 16);
 
     GX_WRITE_U8 (0x10);
     GX_WRITE_U32 (reg);
@@ -410,9 +409,9 @@ GXLoadTexMtxIndx (u16 mtx_indx, u32 id, GXTexMtxType type)
     {
         offset = id * 4;
     }
-    count = (type == GX_MTX2x4) ? 8 : 12;
+    count = (u32)((type == GX_MTX2x4) ? 8 : 12);
 
-    reg = 0;
+    reg   = 0;
     SET_REG_FIELD (0x31B, reg, 12, 0, offset);
     SET_REG_FIELD (0x31C, reg, 4, 12, (count - 1));
     SET_REG_FIELD_ (0x31D, reg, 16, 16, mtx_indx);
@@ -442,20 +441,20 @@ GXSetViewportJitter (f32 left, f32 top, f32 wd, f32 ht, f32 nearz, f32 farz, u32
     {
         top -= 0.5f;
     }
-    sx = wd / 2.0f;
-    sy = -ht / 2.0f;
-    ox = 340.0f + (left + (wd / 2.0f));
-    oy = 340.0f + (top + (ht / 2.0f));
-    zmin = 1.6777215e7f * nearz;
-    zmax = 1.6777215e7f * farz;
-    sz = zmax - zmin;
-    oz = zmax;
-    __GXData->vpLeft = left;
-    __GXData->vpTop = top;
-    __GXData->vpWd = wd;
-    __GXData->vpHt = ht;
+    sx                = wd / 2.0f;
+    sy                = -ht / 2.0f;
+    ox                = 340.0f + (left + (wd / 2.0f));
+    oy                = 340.0f + (top + (ht / 2.0f));
+    zmin              = 1.6777215e7f * nearz;
+    zmax              = 1.6777215e7f * farz;
+    sz                = zmax - zmin;
+    oz                = zmax;
+    __GXData->vpLeft  = left;
+    __GXData->vpTop   = top;
+    __GXData->vpWd    = wd;
+    __GXData->vpHt    = ht;
     __GXData->vpNearz = nearz;
-    __GXData->vpFarz = farz;
+    __GXData->vpFarz  = farz;
     if (__GXData->fgRange != 0)
     {
         __GXSetRange (nearz, __GXData->fgSideX);
@@ -531,15 +530,15 @@ GXGetScissor (u32* left, u32* top, u32* wd, u32* ht)
 
     ASSERTMSGLINE (0x3DD, left && top && wd && ht, "GXGet*: invalid null pointer");
 
-    tp = __GXData->suScis0 & 0x7FF;
-    lf = (__GXData->suScis0 & 0x7FF000) >> 12;
-    bm = __GXData->suScis1 & 0x7FF;
-    rt = (__GXData->suScis1 & 0x7FF000) >> 12;
+    tp    = __GXData->suScis0 & 0x7FF;
+    lf    = (__GXData->suScis0 & 0x7FF000) >> 12;
+    bm    = __GXData->suScis1 & 0x7FF;
+    rt    = (__GXData->suScis1 & 0x7FF000) >> 12;
 
     *left = lf - 340;
-    *top = tp - 340;
-    *wd = rt - lf + 1;
-    *ht = bm - tp + 1;
+    *top  = tp - 340;
+    *wd   = rt - lf + 1;
+    *ht   = bm - tp + 1;
 }
 
 void
